restructure keyboard client code to allow easier reuse (bsc#1174856)
